+++ enigma/cogs/fun.py
@@ -13,7 +13,6 @@ from praw import Reddit
 from requests import request
 
 from enigma.settings import reddit_settings, in_production, rapidapi_settings
-from enigma.utils.colors import random_color
 from enigma.utils.database import create_giveaway, get_giveaway_from_message, delete_giveaway
 from enigma.utils.emebds.core import ErrorEmbed, InfoEmbed, SuccessEmbed
 from enigma.utils.emebds.errors import TimeoutEmbed
@@ -49,8 +48,7 @@ class Fun(Cog):
         if not option:
             await ctx.send(embed=ErrorEmbed(
                 author=ctx.author,
-                title=':x: Please specify an option',
-                color=random_color()
+                title=':x: Please specify an option'
             ))
         else:
             def check(m):
@@ -201,7 +199,6 @@ class Fun(Cog):
 
                         # Preparing final giveaway message
                         final_em.title = ':gift: Giveaway!'
-                        final_em.color = random_color()
                         final_em.add_field(
                             name='\u200b',
                             value='React with `📝` to participate!',