+++ enigma/cogs/admin.py
@@ -3,9 +3,9 @@ from discord import Member, NotFound, HTTPException, User
 from discord.ext.commands import command, Cog, has_permissions, bot_has_permissions, Context, cooldown, BucketType
 from discord.utils import get
 
-from enigma.settings import in_production
 from enigma.emebds.core import ErrorEmbed, SuccessEmbed
 from enigma.emebds.misc import PleaseWaitEmbed
+from enigma.settings import in_production
 
 
 class Admin(Cog):