+++ enigma/settings/database.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from os import environ as env
+
+
+def database_url() -> str:
+    return env.get('DATABASE_URL')